Fix the dependency to the CommandLine.NET issue
The only reason the analyzer has for referencing the actual product is so that the tests
can resolve the types coming from the assembly.

Adding a reference to an older version of the package solves this dependency.

It will need to be bootstrapped every time a new type is added to the package.