Make sure that TryParse and Parse throw when help is requested.
This ensures that when the help is requested we don't return 'true' and give back a null options object.
With this approach we are going to throw a new exception HelpRequestedException for the case when we tried to parse the arguments and they turned out to be a help request.